Reverted op_sequence. Made changes to op_range instead.
 - Made op_range work with a start value,begin_index,end_index instead.
   Makes it more sliceable.
