+++ oss_src/sframe_query_engine/operators/operator_properties.cpp
@@ -52,8 +52,6 @@ RetType extract_field(planner_node_type ptype, CallArgs... call_args) {
       return FieldExtractionVisitor<planner_node_type::GENERALIZED_UNION_PROJECT_NODE>::get(call_args...);
     case planner_node_type::IDENTITY_NODE:
       return FieldExtractionVisitor<planner_node_type::IDENTITY_NODE>::get(call_args...);
-    case planner_node_type::SEQUENCE_NODE:
-      return FieldExtractionVisitor<planner_node_type::SEQUENCE_NODE>::get(call_args...);
     case planner_node_type::INVALID:
       ASSERT_MSG(false, "Infering type of an invalid node");
       return RetType();
