+++ oss_src/sframe_query_engine/operators/range.hpp
@@ -32,46 +32,54 @@ struct operator_impl<planner_node_type::RANGE_NODE> : public query_operator {
     return ret;
   }
   
-  inline operator_impl(flex_int begin_index, flex_int end_index) { ASSERT_LE(begin_index, end_index); };
-  
-  inline std::string print() const {
-    return name() + "(" + std::to_string(m_begin_index) + ", " + std::to_string(m_end_index) + ")";
+  inline operator_impl(flex_int start, flex_int end)
+  : m_start(start)
+  , m_end(end)
+  { 
+    ASSERT_LE(m_start, m_end);
   }
-
+  
+  
   inline std::shared_ptr<query_operator> clone() const {
-    return std::make_shared<operator_impl>(*this);
+    return std::make_shared<operator_impl>(m_start, m_end);
   }
 
   inline void execute(query_context& context) {
-    ASSERT_LE(m_begin_index, m_end_index);
-    flex_int iter = m_begin_index;
-    while(iter != m_end_index) {
+    flex_int cur = m_start;
+
+    while(cur < m_end) {
       auto ret = context.get_output_buffer();
-      size_t len = std::min<size_t>(m_end_index - iter, context.block_size());
+      size_t len = std::min<size_t>(m_end - cur, context.block_size());
 
       ret->resize(1, len);
       for (auto& value: *(ret->get_columns()[0])) {
-        value = iter;
-        ++iter;
+        value = cur;
+        ++cur;
       }
       context.emit(ret);
     }
-  };
+  }
 
   static std::shared_ptr<planner_node> make_planner_node(
-      flex_int begin_index, flex_int end_index) {
+      flex_int start, flex_int end) {
     return planner_node::make_shared(planner_node_type::RANGE_NODE, 
-                                     {{"begin_index", begin_index},
-                                       {"end_index", end_index}});
+                                     {{"start", start},
+                                      {"begin_index", 0},
+                                      {"end_index", end - start}});
   }
+
   static std::shared_ptr<query_operator> from_planner_node(
       std::shared_ptr<planner_node> pnode) {
-        ASSERT_EQ((int)pnode->operator_type, 
-                  (int)planner_node_type::RANGE_NODE);
-        ASSERT_TRUE(pnode->operator_parameters.count("begin_index"));
-        ASSERT_TRUE(pnode->operator_parameters.count("end_index"));
-        return std::make_shared<operator_impl>(pnode->operator_parameters["begin_index"],
-                                               pnode->operator_parameters["end_index"]);
+    ASSERT_EQ((int)pnode->operator_type, (int)planner_node_type::RANGE_NODE);
+    ASSERT_TRUE(pnode->operator_parameters.count("start"));
+    ASSERT_TRUE(pnode->operator_parameters.count("begin_index"));
+    ASSERT_TRUE(pnode->operator_parameters.count("end_index"));
+
+    flex_int start = (flex_int)pnode->operator_parameters["start"];
+    size_t begin_index = pnode->operator_parameters["begin_index"];
+    size_t end_index = pnode->operator_parameters["end_index"];
+    return std::make_shared<operator_impl>(start + begin_index, 
+                                           start + end_index);
   }
 
   static std::vector<flex_type_enum> infer_type(
@@ -80,30 +88,35 @@ struct operator_impl<planner_node_type::RANGE_NODE> : public query_operator {
     return {flex_type_enum::INTEGER};
   }
 
+
   static int64_t infer_length(std::shared_ptr<planner_node> pnode) {
     ASSERT_EQ((int)pnode->operator_type, (int)planner_node_type::RANGE_NODE);
-
-    flex_int begin_index = pnode->operator_parameters.at("begin_index");
-    flex_int end_index = pnode->operator_parameters.at("end_index");
-
-    return end_index - begin_index;
+    size_t count = pnode->operator_parameters["end_index"] -
+                   pnode->operator_parameters["begin_index"];
+    return count;
   }
-
+  
   static std::string repr(std::shared_ptr<planner_node> pnode, pnode_tagger&) {
+    ASSERT_EQ((int)pnode->operator_type, (int)planner_node_type::RANGE_NODE);
+    ASSERT_TRUE(pnode->operator_parameters.count("start"));
     ASSERT_TRUE(pnode->operator_parameters.count("begin_index"));
     ASSERT_TRUE(pnode->operator_parameters.count("end_index"));
 
+    flex_int start = pnode->operator_parameters["start"];
     size_t begin_index = pnode->operator_parameters["begin_index"];
     size_t end_index = pnode->operator_parameters["end_index"];
 
-    std::ostringstream out;
-    out << "Range(" << begin_index << ',' << end_index << ")";
-    return out.str();
+    std::ostringstream ss;
+
+    ss << "Sequence(" << start << ")[" << begin_index << ":" << end_index << "]";
+
+    return ss.str();
   }
 
  private:
-  flex_int m_begin_index;
-  flex_int m_end_index;
+  // m_start to m_end defines the range
+  flex_int m_start; // inclusive
+  flex_int m_end;   // exclusive
 };
 
 typedef operator_impl<planner_node_type::RANGE_NODE> op_range;
